pyproject.toml
[project]
name = "fastapi-project"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.9"
dependencies = [
    "aiosqlite>=0.21.0",
    "fastapi>=0.124.2",
    "fastapi-users[sqlalchemy]>=14.0.2",
    "imagekitio>=4.2.0",
    "python-dotenv>=1.2.1",
    "streamlit>=1.50.0",
    "uvicorn[standard]>=0.38.0",
]
